% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{get_version}
\alias{get_version}
\title{Determine API endpoint version}
\usage{
get_version(api = c("nlp", "tts", "speech", "trans"))
}
\arguments{
\item{api}{Character, one of "nlp", "tts", "speech", "trans"}
}
\value{
Character string of API version
}
\description{
Used to specify the API endpoint: GitHub beta versions or CRAN stable.
}
\keyword{internal}
