% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural-language.R
\name{gl_nlp_entity}
\alias{gl_nlp_entity}
\title{Get Entities from a string}
\usage{
gl_nlp_entity(string, type = c("PLAIN_TEXT", "HTML"), language = c("en",
  "es", "ja"))
}
\arguments{
\item{string}{A character vector of text to detect language for}

\item{type}{Whether its text or a HTML page}

\item{language}{Language of source, must be supported.}
}
\value{
A list of entities of the text
}
\description{
Get Entities from a string
}
\seealso{
\url{https://cloud.google.com/natural-language/docs/reference/rest/v1/documents/analyzeEntities}
}
