% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech.R
\name{gl_speech_record}
\alias{gl_speech_record}
\title{Record audio from an R session}
\usage{
gl_speech_record(filename = "inst/test.wav", sampleRateHertz = 16000L,
  secs = 15L)
}
\arguments{
\item{filename}{Where to save WAV file as uncompressed PCM data}

\item{sampleRateHertz}{Sample rate for the recording}

\item{secs}{How long to record for

Play the file using the same sampleRate e.g. \code{audio::play(a$data, rate = 16000L)}}
}
\value{
The audio data of class \code{audioSample}
}
\description{
Wrapper around \link[audio]{record} with defaults that can be used in \link{gl_speech_recognition}
}
