% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{gl_translate_language}
\alias{gl_translate_language}
\title{Translate the language of text within a request}
\usage{
gl_translate_language(string, encode = TRUE, target = "en",
  format = c("text", "html"), source = "", model = c("nmt", "base"))
}
\arguments{
\item{string}{A character vector of text to detect language for}

\item{encode}{If TRUE, will run strings through URL encoding}

\item{target}{The target language}

\item{format}{Whether the text is plain or HTML}

\item{source}{Specify the language to translate from. Will detect it if left default}

\item{model}{What translation model to use}
}
\value{
A list of the detected languages
}
\description{
Translate the language of text within a request
}
\seealso{
\url{https://cloud.google.com/translate/docs/reference/translate}

Other translations: \code{\link{gl_translate_detect}},
  \code{\link{gl_translate_list}}
}
