% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{gl_translate_list}
\alias{gl_translate_list}
\title{List languages from Google Translate API}
\usage{
gl_translate_list(target = "en")
}
\arguments{
\item{target}{If specified, language names are localized in target langauge}
}
\value{
data.frame of supported languages
}
\description{
Returns a list of supported languages for translation.
}
\details{
Supported language codes, generally consisting of its ISO 639-1 identifier. (E.g. \code{'en', 'ja'}).
In certain cases, BCP-47 codes including language + region identifiers are returned (e.g. \code{'zh-TW', 'zh-CH'})
}
\seealso{
\url{https://cloud.google.com/translate/docs/reference/languages}

Other translations: \code{\link{gl_translate_detect}},
  \code{\link{gl_translate_language}}
}
