% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleLanguageR.R
\docType{package}
\name{googleLanguageR-package}
\alias{googleLanguageR}
\alias{googleLanguageR-package}
\title{googleLanguageR}
\description{
This package contains functions for analysing language through the
  Google Cloud Machine Learning APIs
}
\details{
For examples and documentation see the vignettes and the website:

\url{https://github.com/ropensci/googleLanguageR}
}
\seealso{
\url{https://cloud.google.com/products/machine-learning}

googleLanguageR: Interface to Google Cloud NLP, Translation, and Speech APIs

A package for interacting with Google Cloud's language APIs from R.
}
\author{
\strong{Maintainer}: Cheryl Isabella \email{cheryl.academic@gmail.com}

Authors:
\itemize{
  \item Mark Edmondson \email{r@sunholo.com}
}

Other contributors:
\itemize{
  \item Aleksander Dietrichson \email{dietrichson@gmail.com} [contributor]
  \item John Muschelli \email{muschellij2@gmail.com} [contributor]
  \item Neal Richardson \email{neal.p.richardson@gmail.com} (Neal reviewed the package for ropensci) [reviewer]
  \item Julia Gustavsen \email{j.gustavsen@gmail.com} (Julia reviewed the package for ropensci) [reviewer]
}
}
