% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleLanguageR.R
\docType{package}
\name{googleLanguageR}
\alias{googleLanguageR}
\alias{googleLanguageR-package}
\title{googleLanguageR}
\description{
This package contains functions for analysing language through the
  Google Cloud Machine Learning APIs
}
\details{
For examples and documentation see the vignettes and the website:

\url{http://code.markedmondson.me/googleLanguageR/}
}
\seealso{
\url{https://cloud.google.com/products/machine-learning/}
}
