% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is.gcs}
\alias{is.gcs}
\title{Test if a string is a Google Cloud Storage URI}
\usage{
is.gcs(x)
}
\arguments{
\item{x}{Character string}
}
\value{
Logical
}
\description{
Test if a string is a Google Cloud Storage URI
}
\keyword{internal}
