% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is.NullOb}
\alias{is.NullOb}
\title{A helper function that tests whether an object is either NULL _or_
a list of NULLs}
\usage{
is.NullOb(x)
}
\description{
A helper function that tests whether an object is either NULL _or_
a list of NULLs
}
\keyword{internal}
