% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{my_message}
\alias{my_message}
\title{Custom messaging function for package}
\usage{
my_message(..., level = 1)
}
\arguments{
\item{...}{Messages to print}

\item{level}{Message level (1=low, 2=debug, 3=normal)}
}
\description{
Custom messaging function for package
}
\keyword{internal}
