% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{gl_auth}
\alias{gl_auth}
\title{Auth with Google Translate API}
\usage{
gl_auth(json_file)
}
\arguments{
\item{json_file}{File you have downloaded from your Google Project}
}
\description{
Auth with Google Translate API
}
