% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural-language.R
\name{gl_nlp}
\alias{gl_nlp}
\title{Perform Natural Language Analysis on text}
\usage{
gl_nlp(string, nlp_type = c("annotateText", "analyzeEntities",
  "analyzeSentiment", "analyzeSyntax", "analyzeEntitySentiment"),
  type = c("PLAIN_TEXT", "HTML"), language = c("en", "zh", "zh-Hant", "fr",
  "de", "it", "ja", "ko", "pt", "es", "de"), encodingType = c("UTF8", "UTF16",
  "UTF32", "NONE"), version = c("v1", "v1beta2", "v1beta1"))
}
\arguments{
\item{string}{A character vector of text to detect language for, or Google Cloud Storage URIs}

\item{nlp_type}{The type of Natural Language Analysis to perform.  The default \code{annotateText} will perform all features in one call.}

\item{type}{Whether input text is plain text or a HTML page}

\item{language}{Language of source, must be supported by API.  If \code{NULL} is auto-detected.}

\item{encodingType}{Text encoding that the caller uses to process the output}

\item{version}{the API version}
}
\value{
A list of analysis of the text
}
\description{
Analyse text entites, sentiment, and syntax using the Google Natural Language API
}
\details{
\code{string} can be a character vector, or a location of a file content on Google cloud Storage.
  This URI must be of the form \code{gs://bucket_name/object_name}

Encoding type can usually be left at default \code{UTF8}.
  See here for details: \url{https://cloud.google.com/natural-language/docs/reference/rest/v1/EncodingType}

The current language support is available here \url{https://cloud.google.com/natural-language/docs/languages}
}
\examples{

\dontrun{

text <- "to administer medicince to animals is frequently a very difficult matter,
  and yet sometimes it's necessary to do so"
nlp_result <- gl_nlp(text)

head(nlp_result$tokens$text)

head(nlp_result$tokens$partOfSpeech$tag)

nlp_result$entities

nlp_result$documentSentiment

}

}
\seealso{
\url{https://cloud.google.com/natural-language/docs/reference/rest/v1/documents}
}
