% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-to-speech.R
\name{gl_talk_languages}
\alias{gl_talk_languages}
\title{Get a list of voices available for text to speech}
\usage{
gl_talk_languages(languageCode = NULL)
}
\arguments{
\item{languageCode}{A \code{BCP-47} language tag.  If specified, will only return voices that can be used to synthesize this languageCode}
}
\description{
Returns a list of voices supported for synthesis.
}
