% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-to-speech.R
\name{gl_talk_player}
\alias{gl_talk_player}
\title{Play audio in a browser}
\usage{
gl_talk_player(audio = "output.wav", html = "player.html")
}
\arguments{
\item{audio}{The file location of the audio file.  Must be supported by HTML5}

\item{html}{The html file location that will be created host the audio}
}
\description{
This uses HTML5 audio tags to play audio in your browser
}
\details{
A platform neutral way to play audio is not easy, so this uses your browser to play it instead.
}
\examples{

\dontrun{

gl_talk("Testing my new audio player") \%>\% gl_talk_player()

}

}
