% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-to-speech.R
\name{gl_talk_shinyUI}
\alias{gl_talk_shinyUI}
\title{Speak in Shiny module (ui)}
\usage{
gl_talk_shinyUI(id)
}
\arguments{
\item{id}{The Shiny id}
}
\description{
Speak in Shiny module (ui)
}
\details{
Shiny Module for use with \link{gl_talk_shiny}.
}
