% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{gl_translate_detect}
\alias{gl_translate_detect}
\title{Detect the language of text within a request}
\usage{
gl_translate_detect(string, encode = TRUE)
}
\arguments{
\item{string}{A character vector of text to detect language for}

\item{encode}{If TRUE, will run strings through URL encoding}
}
\value{
A list of the detected languages
}
\description{
Detect the language of text within a request
}
\details{
Consider using \code{library(cld2)} and \code{cld2::detect_language} instead offline,
since that is free and local without needing a paid API call.

\link{gl_translate_language} also returns a detection of the language,
so you could also wish to do it in one step via that function.
}
\examples{

\dontrun{
## which language is this?
gl_translate_detect("katten sad på måtten")
# confidence isReliable language
#1  0.1863063      FALSE       sv

gl_translate_detect("katten sidder på måtten")
# Detecting language: 39 characters - katten sidder på måtten...
# [[1]]
# isReliable language confidence
# 1      FALSE       da   0.536223


}

}
\seealso{
\url{https://cloud.google.com/translate/docs/reference/detect}

Other translations: \code{\link{gl_translate_language}},
  \code{\link{gl_translate_list}}
}
