% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate-documents.R
\name{gl_translate_document}
\alias{gl_translate_document}
\title{Translate document}
\usage{
gl_translate_document(
  d_path,
  target = "en",
  format = c("pdf"),
  source = "",
  model = c("nmt", "base")
)
}
\arguments{
\item{d_path}{path to the document}

\item{target}{The target language}

\item{format}{Whether the text is plain or HTML}

\item{source}{Specify the language to translate from. Will detect it if left default}

\item{model}{What translation model to use}
}
\description{
Translate a document via the Google Translate API
}
