% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate-document.R
\name{gl_translate_document}
\alias{gl_translate_document}
\title{Translate document}
\usage{
gl_translate_document(
  d_path,
  target = "es-ES",
  output_path = "out.pdf",
  format = c("pdf"),
  source = "en-UK",
  model = c("nmt", "base"),
  location = "global"
)
}
\arguments{
\item{d_path}{path of the document to be translated}

\item{output_path}{where to save the translated document}

\item{format}{currently only pdf-files are supported}
}
\value{
output filename
}
\description{
Translate a document via the Google Translate API
}
\examples{

\dontrun{
gl_translate_document(system.file(package = "googleLanguageR","test-doc.pdf"), "no")

}
}
\seealso{
Other translations: 
\code{\link{gl_translate_detect}()},
\code{\link{gl_translate_languages}()},
\code{\link{gl_translate}()}
}
\concept{translations}
