% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{gl_translate_language}
\alias{gl_translate_language}
\title{Translate the language of text within a request}
\usage{
gl_translate_language(string, encode = TRUE, target = "en",
  format = c("text", "html"), source = "", model = c("nmt", "base"))
}
\arguments{
\item{string}{A character vector of text to detect language for}

\item{encode}{If TRUE, will run strings through URL encoding}

\item{target}{The target language}

\item{format}{Whether the text is plain or HTML}

\item{source}{Specify the language to translate from. Will detect it if left default}

\item{model}{What translation model to use}
}
\value{
A list of the translations
}
\description{
Translate character vectors via the Google Translate API
}
\details{
You can translate a vector of strings, but each call must be under 2000 characters.
If more than 2000 characters, split it up into seperate calls.

The API limits in three ways: characters per day, characters per 100 seconds, and API requests per 100 seconds. All can be set in the API manager \code{https://console.developers.google.com/apis/api/translate.googleapis.com/quotas}

The function will limit the API calls for for the characters and API requests per 100 seconds, which you can set via the options \code{googleLanguageR.rate_limit} and \code{googleLanguageR.character_limit}.  By default these are set at \code{0.5} requests per second, and \code{100000} characters per 100 seconds.
}
\examples{

\dontrun{

text <- "to administer medicince to animals is frequently a very difficult matter,
  and yet sometimes it's necessary to do so"

japan <- gl_translate_language(text, target = "ja")

japan$translatedText
#> [1] "動物に薬を投与することはしばしば非常に困難な問題ですが、時にはそれを行う必要があります"


}

}
\seealso{
\url{https://cloud.google.com/translate/docs/reference/translate}

Other translations: \code{\link{gl_translate_detect}},
  \code{\link{gl_translate_list}}
}
